# openapi3.yaml
openapi: 3.0.1
info:
  title: Discrete-agent-DB
  description: basic template for map colonies service
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /status:
    get:
      operationId: getStatus
      tags:
        - Status
      summary: gets the agent status
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status'
    put:
      operationId: updateStatus
      tags:
        - Status
      summary: updates the agent status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/status'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/status'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /layers/{directory}:
    get:
      operationId: getLayerHistory
      tags:
        - LayerHistory
      summary: gets the history record of specific layer
      parameters:
        - $ref: '#/components/parameters/directoryParam'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/layerHistoryRecord'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    post:
      operationId: CreateLayerHistory
      tags:
        - LayerHistory
      summary: creates history record for specific layer
      parameters:
        - $ref: '#/components/parameters/directoryParam'
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/layerHistoryRecord'
        409:
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
    put:
      operationId: UpdateLayerHistoryStatus
      tags:
        - LayerHistory
      summary: update the status of history record for specific layer
      parameters:
        - $ref: '#/components/parameters/directoryParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                version:
                  type: string
                status:
                  $ref: '#/components/schemas/historyStatus'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/layerHistoryRecord'
        404:
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
components:
  parameters:
    directoryParam:
      name: directory
      in: path
      description: layer source directory
      required: true
      schema:
        type: string
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    status:
      type: object
      required:
        - isWatching
      properties:
        isWatching:
          type: boolean
    layerHistoryRecord:
      type: object
      properties:
        directory:
          type: string
        id:
          type: string
        version:
          type: string
        status:
          $ref: '#/components/schemas/historyStatus'
    historyStatus:
      type: string
      enum:
        - inProgress
        - triggered
        - failed